try not packing any fonts
fonts are a sizable chunk of the snap size, but for quite a while now
the desktop/unity7/x11 interfaces should provide access to the system
fonts so we don't need to ship our own. see if this actually works.

the content snap goes into candidate channel by default, so testing this
does no harm to production