implement proper versioning of interfaces
currently we have the problem that we repack debs into snaps which means
they were built against say Qt 5.10 while the content snap can still be
Qt 5.9. this specifically also applies to the users system as they might
refresh an app but not the framework snap which means the dependency is
now broken as the correct Qt can't be found by ld until the content snap
is also updated

since content snaps have no versioning system we'll instead have to use
specific interfaces. so for example okular needs to require
kde-frameworks-5-46-qt-5-10-ubuntu-1604 so it correctly loads the binary
compatible framework content snap

to facilitiate this the atomizer now does a number of new things
- versions.json is used to track ALL previous full versions of the content
  snap
- this is archived/unarchived by the Jenkinsfile
- the versions.json is loaded in the atomizer and used to auto-create
  interface. they are essentially aliases for one another
- upon iterating the parts the versions of kf5 and qt are extracted and
  the CURRENT version is constructed
- the current version is dumped into a content.json file so app builds can
  read that to figure out the latest version, this would be the version
  they need to require so as to stay binary compatible (this is
  declarative; since the snap builds are largely repacks of debs they are
  already binary requiring the latest kf5/qt5, this simply declares the
  fact to snapd)
- if the current version isn't yet tracked in version.json it is added
  and the versions.json is dumped to file again

ideally versions.json should live in git and content.json only in the
workspace. unfortunately pushing to git is proofing a bit difficult as
the repo is still on github and we have no bot account there (also
since pipeline's git has no builtin push support we'd have to handle it
via scripting, specifically also we'd need to handle credentials which
is a bit of a mess)