do not link snapcraft to generate target
unfortunately due to how we build on the jenkins side rake will always
think the file targets are out of date (timestamps suggest it needs
rebuilding bc neon.list is newer) so we'd incorrectly attempt to generate
during the snapcraft part of the jenkins job.
to prevent this we go back to where we were before. two targets without
interaction