exhaustively deal with qtchooser
the amount of time that tool has cost everyone by now must be in the
hundred of hours -.-

qtchooser determines the exe it needs to launch by checking its own exename
so the qname ->symlink-> qtchooser trick results in qtchooser knowing
that it needs to lunch qname.
our previous wrapper tech simply moved qname to qname.orig and exec'd that
which of course broke qtchooser.
new tech instead moves the qname.orig to a subdir so we don't need the
orig.
this should allow wrapping to work as before but also for qtchooser